+++ src/chains/Taiko.sol
@@ -20,6 +20,8 @@ import {
 } from "../core/univ3forks/UniswapV3.sol";
 import {swapsicleFactory, swapsicleInitHash, swapsicleForkId} from "../core/univ3forks/Swapsicle.sol";
 import {IAlgebraCallback} from "../core/univ3forks/Algebra.sol";
+import {pankoFactory, pankoInitHash, pankoForkId} from "../core/univ3forks/Panko.sol";
+import {IPancakeSwapV3Callback} from "../core/univ3forks/PancakeSwapV3.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../SettlerAbstract.sol";
@@ -60,6 +62,10 @@ abstract contract TaikoMixin is FreeMemory, SettlerBase {
             factory = swapsicleFactory;
             initHash = swapsicleInitHash;
             callbackSelector = uint32(IAlgebraCallback.algebraSwapCallback.selector);
+        } else if (forkId == pankoForkId) {
+            factory = pankoFactory;
+            initHash = pankoInitHash;
+            callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
         } else {
             revert UnknownForkId(forkId);
         }