+++ CHANGELOG.md
@@ -24,7 +24,8 @@ Master list of UniV3 forks:
   21. SupSwapV3
   22. Kim (Algebra-like)
   23. SwapMode
-  24. Swapsicle
+  24. Swapsicle (Algebra-like)
+  25. Panko
 
 ---
 
@@ -55,6 +56,7 @@ Master list of UniV3 forks:
 * Deploy Settler to Taiko network
   * Add UniswapV3 UniV3 fork to Taiko
   * Add Swapsicle UniV3 (Algebra style) fork to Taiko
+  * Add Panko UniV3 fork to Taiko
 
 ## 2024-10-08